changes: removed react-query and added redux toolkit; feature: search more button
at first, it was just meant to be the feature, but I was having trouble trying to make it clean, and code would get really messy, so I decided to switch to redux, because of it's way to manage states (I needed information to be pushed to an array, but useState just changes it as
a whole, even if I do it like setState([...state, ...newInfo]). Redux lets me push the info, so
I always have it... or that's what I thought. In the end, I realized the problem was how I was trying to render the cards. I was doing it conditionally (!isLoading && !isError) then render the cards, at what it did is that it made my cards
disappear for almost 2 seconds. What I should've done is just always render them, but it would've been a problem since at first, when the page loads, I don't have any, and code would get really messy. With redux it's just simpler and easier and cleaner (at least that's how I see it)