change: data fetched
before I was fetching the whole pokemon data, and twice, so my page was really slow. But I discovered that I can get the sprite of the pokemon by just changing a number in the link, and I don't really need the id, I can make one using the index of the array of pokemon. So, I switched from fetching a whole pokemon and pokemon-species data, and just fetch once the list of pokemon, then pass the link for the sprite to the component