feature: search pokemon variants
I had to first check if the user searched for a name or id, then if it was a name, check if it has spaces on it and replace them with '-', then check if it says 'alolan' or 'galarian' and replace them with 'alola' and 'galar', then fetch the pokemon that way, if it fails, it means two things: the user wrote a wrong nam/id, or they are looking for a variant, so I check the latter, reverting the order (because pokemon-species doesn't have variants). If it still fails, it means that they wrote the name backwards (e.g. 'alola raichu' instead of 'raichu alola', the variant must always go second), so I reverse the name and search again. If it still fails, it means the user did in fact write a wrong name/id