refactored the singlePoke fetching
now it uses fewer lines of code, does only one fetching, and the user can make an incomplete search as long as it gives something. E.g.: the user is searching 'pik ro', and the fetching will be 'pikachu-rock-star'. This is because I first fetch the entire pokemon list on page load (which is nothing actually, 1118 poke, and it's just the names), so I match the user's input with the base I already have. If what they wrote matches at least one of the poke, then I fetch that. And because I'm using includes() instead of an exact match, I can fetch an incomplete name